Moved gift scheduler endpoints into one controller
ref https://linear.app/ghost/issue/BER-3851/establish-immediate-email-delivery-for-gift-subscriptions

Reminder and delivery wake-ups share the same gifts adapter, so keeping them together reduces duplicate endpoint plumbing without changing their routes or permissions.
