+++ ghost/core/core/server/api/endpoints/index.js
@@ -316,12 +316,8 @@ module.exports = {
         return apiFramework.pipeline(require('./gift-links'), localUtils);
     },
 
-    get giftReminders() {
-        return apiFramework.pipeline(require('./gift-reminders'), localUtils);
-    },
-
-    get giftDeliveries() {
-        return apiFramework.pipeline(require('./gift-deliveries'), localUtils);
+    get gifts() {
+        return apiFramework.pipeline(require('./gifts'), localUtils);
     },
 
     get recommendationsPublic() {
